rm fill_tx_env from primitives
