+++ examples/exex/rollup/src/execution.rs
@@ -10,7 +10,6 @@ use reth_primitives::{
     constants,
     eip4844::kzg_to_versioned_hash,
     keccak256,
-    revm::env::fill_tx_env,
     revm_primitives::{CfgEnvWithHandlerCfg, EVMError, ExecutionResult, ResultAndState},
     Address, Block, BlockWithSenders, Bytes, EthereumHardfork, Header, Receipt, TransactionSigned,
     TxType, B256, U256,
@@ -217,7 +216,7 @@ fn execute_transactions(
             }
             // Execute transaction.
             // Fill revm structure.
-            fill_tx_env(evm.tx_mut(), &transaction, sender);
+            EthEvmConfig::default().fill_tx_env(evm.tx_mut(), &transaction, sender);
 
             let ResultAndState { result, state } = match evm.transact() {
                 Ok(result) => result,
